Support multiple environments directories
This commit add the support for multiple environments directories, like having:

('/home/user/.conda/envs', '/usr/envs')

For achieving this, some other enhancements have been done:

- It checks for duplicated environment names, and in this case raises an error
saying that this is not supported.

- The $PATH management has also been simplified for supporting multiple
environments directories: When activating an environment, the *system's* bin dir
is not replaced, but instead the *environment's* bin dir is placed just before
it. When deactivating an environment, the environment's bin dir is simply
removed from $PATH.

Signed-off-by: Fabien Dubosson <fabien.dubosson@gmail.com>