Remove $SHELL from `symlink_conda` call
`$SHELL` should return the users default shell, which might not be `xonsh`, so
this shouldn't be in there anyway.
In any case, it seems like the `shell=` kwarg is only used on Windows and even
there I don't know if it's required.

If it turns out to be an issue for windows users, I'm sure they'll let me know.