Fully remove current conda bin from path on activate
This functionality was the original behavior of `xonda` and was removed in a
later commit that added support for multiple env_dirs because the bookkeeping
was complicated.  This has become enough of a pain point that I am putting it
back in -- if you activate an environment and type `ipython` and you don't have
`ipython` installed in the activated environment, that should error out.  But it
doesn't with the current (pre this commit) behavior, because it falls back on
the `ipython` installed in the default anaconda env.

That is the behavior of `activate` that is shipped with Anaconda when used in
the `bash` or `zsh` world with `source activate env_name` but it is a terrible
behavior.